+++ app/src/main/res/layout/activity_main.xml
@@ -1,17 +1,9 @@
-<LinearLayout
+<FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_height="fill_parent"
     android:layout_width="fill_parent"
     android:id="@+id/grid_parent"
+    android:background="@color/background_material_dark"
     android:padding="@dimen/default_border_padding">
 
-    <GridView
-        android:id="@+id/image_grid"
-        android:layout_width="match_parent"
-        android:numColumns="3"
-        android:verticalSpacing="@dimen/grid_row_vertical_spacing"
-        android:horizontalSpacing="@dimen/grid_row_vertical_spacing"
-        android:layout_height="match_parent">
-    </GridView>
-
-</LinearLayout>
+</FrameLayout>
