+++ app/src/main/res/layout/grid_frag_layout.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<GridView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/image_grid"
+    android:layout_width="fill_parent"
+    android:numColumns="3"
+    android:verticalSpacing="@dimen/grid_row_vertical_spacing"
+    android:horizontalSpacing="@dimen/grid_row_vertical_spacing"
+    android:layout_height="fill_parent">
+</GridView>
